Add NewInstaller method to client interface
