+++ pkg/helmcli/install.go
@@ -8,32 +8,14 @@ import (
 	"helm.sh/helm/v3/pkg/chart/loader"
 	"helm.sh/helm/v3/pkg/cli"
 	"helm.sh/helm/v3/pkg/release"
-
-	"github.com/gojekfarm/albatross/pkg/helmcli/config"
-	"github.com/gojekfarm/albatross/pkg/helmcli/flags"
 )
 
-type Installer struct {
+type installer struct {
 	action      *action.Install
 	envSettings *cli.EnvSettings
 }
 
-// NewInstaller returns a new instance of Installer struct
-func NewInstaller(flg flags.InstallFlags) *Installer {
-	envconfig := config.NewEnvConfig(&flg.GlobalFlags)
-	actionconfig := config.NewActionConfig(envconfig, &flg.GlobalFlags)
-
-	install := action.NewInstall(actionconfig.Configuration)
-	install.Namespace = flg.Namespace
-	install.DryRun = flg.DryRun
-
-	return &Installer{
-		action:      install,
-		envSettings: envconfig.EnvSettings,
-	}
-}
-
-func (i *Installer) Install(ctx context.Context, relName, chartName string, values map[string]interface{}) (*release.Release, error) {
+func (i *installer) Install(ctx context.Context, relName, chartName string, values map[string]interface{}) (*release.Release, error) {
 	i.action.ReleaseName = relName
 
 	chart, err := i.loadChart(chartName)
@@ -44,7 +26,7 @@ func (i *Installer) Install(ctx context.Context, relName, chartName string, valu
 	return i.action.Run(chart, values)
 }
 
-func (i *Installer) loadChart(chartName string) (*chart.Chart, error) {
+func (i *installer) loadChart(chartName string) (*chart.Chart, error) {
 	cp, err := i.action.LocateChart(chartName, i.envSettings)
 	if err != nil {
 		return nil, err
